+++ CHANGELOG.md
@@ -1,11 +1,28 @@
-## 0.18.5 (Unreleased)
+## 0.18.6 (Unreleased)
+
+- awsx.ecs.Cluster can be created from an existing aws.ecs.Cluster's id.
+- Add OPTIONS as a valid method and add ability to set custom gateway responses for
+  [awsx.apigateway.API].
+
+### Compatibility issues
+
+- An `awsx.ec2.Vpc` with `assignGeneratedIpv6CidrBlock: true` will now set
+  `assignIpv6AddressOnCreation: true` by default for child subnets.  This can be overridden by
+  setting that value explicitly to `false` with the subnet's args.
+
+
+## 0.18.5 (6/12/2019)
 
 - VPCs can now be made which scale to use all availability zones in a region if desired.  Use
   `new awsx.ec2.Vpc("name", { numberOfAvailabilityZones: "all" })` to get this behavior.  If
   `numberOfAvailabilityZones` is not provided, the current behavior of defaulting to 2 availability
   zones remains.
-- Add OPTIONS as a valid method and add ability to set custom gateway responses for
- [awsx.apigateway.API].
+- Externally available application listeners will now open their security group to both ingress and
+  egress for their specified port.
+- Tweaked API.getFunction to allow [route] and [method] parameters to be optional.  Also changed
+  function to throw if passed arguments that don't map to an actual function.
+- awsx.cloudwatch.Dashboard now exports a `url` property that gives you an immediate link to the
+  Dashboard.
 
 ## 0.18.4 (5/14/2019)